+++ darwin/photo_manager/Sources/photo_manager/core/NSString+PM_COMMON.m
@@ -3,6 +3,9 @@
 @implementation NSString (PM_COMMON)
 
 - (BOOL)isEmpty {
+    if (self == nil || self == (id)[NSNull null]) {
+        return YES;
+    }
     if (self.length == 0) {
         return YES;
     }